Split resolver module graph tests
