+++ src/typechecker/tests/resolver_module_graph/import_signatures.rs
@@ -1,0 +1,79 @@
+use super::super::*;
+
+#[test]
+fn check_module_graph_entry_uses_graph_import_bindings() {
+    let tmp = tempfile::tempdir().expect("temp dir");
+    let math_path = tmp.path().join("math.zen");
+    std::fs::write(&math_path, "pub add = (a: i32, b: i32) i32 { a + b }\n")
+        .expect("write imported module");
+
+    let main_path = tmp.path().join("main.zen");
+    std::fs::write(
+        &main_path,
+        "{ add } = math\n\nmain = () i32 { add(1, 2) }\n",
+    )
+    .expect("write entry module");
+
+    let mut files = crate::error::FileTable::new();
+    let mut modules = crate::module_system::ModuleSystem::new();
+    let graph = modules
+        .load_module_graph(&main_path, &mut files)
+        .expect("module graph");
+    let entry = graph.module(graph.entry).expect("entry module");
+    assert!(
+        !entry
+            .program
+            .declarations
+            .iter()
+            .any(|decl| decl.name() == Some("add")),
+        "graph entry should not merge imported declarations"
+    );
+
+    let mut tc = TypeChecker::new();
+    let typed = tc
+        .check_module_graph_entry(&graph)
+        .expect("graph import bindings should seed imported signatures");
+
+    assert!(typed
+        .functions
+        .iter()
+        .any(|function| function.name == "main"));
+    assert!(typed
+        .functions
+        .iter()
+        .any(|function| function.name == "add"));
+}
+
+#[test]
+fn check_module_graph_entry_seeds_imported_function_type_signatures() {
+    let tmp = tempfile::tempdir().expect("temp dir");
+    let callbacks_path = tmp.path().join("callbacks.zen");
+    std::fs::write(
+        &callbacks_path,
+        "pub apply = (callback: (i32) i32, value: i32) i32 { value }\n",
+    )
+    .expect("write imported module");
+
+    let main_path = tmp.path().join("main.zen");
+    std::fs::write(
+        &main_path,
+        r#"{ apply } = callbacks
+
+main = () i32 {
+    callback = (value: i32) i32 { value }
+    apply(callback, 1)
+}
+"#,
+    )
+    .expect("write entry module");
+
+    let mut files = crate::error::FileTable::new();
+    let mut modules = crate::module_system::ModuleSystem::new();
+    let graph = modules
+        .load_module_graph(&main_path, &mut files)
+        .expect("module graph");
+
+    let mut tc = TypeChecker::new();
+    tc.check_module_graph_entry(&graph)
+        .expect("graph import bindings should seed function-typed signatures");
+}
