+++ src/typechecker/tests/resolver_module_graph/imported_methods.rs
@@ -1,0 +1,130 @@
+use super::super::*;
+
+#[test]
+fn check_module_graph_entry_seeds_public_methods_for_imported_types() {
+    let tmp = tempfile::tempdir().expect("temp dir");
+    let geometry_path = tmp.path().join("geometry.zen");
+    std::fs::write(
+        &geometry_path,
+        r#"pub Point: { x: i32 }
+
+pub Point.value = (self: Point) i32 {
+    self.x
+}
+"#,
+    )
+    .expect("write imported module");
+
+    let main_path = tmp.path().join("main.zen");
+    std::fs::write(
+        &main_path,
+        r#"{ Point } = geometry
+
+main = () i32 {
+    point = Point { x: 7 }
+    point.value()
+}
+"#,
+    )
+    .expect("write entry module");
+
+    let mut files = crate::error::FileTable::new();
+    let mut modules = crate::module_system::ModuleSystem::new();
+    let graph = modules
+        .load_module_graph(&main_path, &mut files)
+        .expect("module graph");
+
+    let mut tc = TypeChecker::new();
+    tc.check_module_graph_entry(&graph)
+        .expect("imported public type should seed its public methods");
+}
+
+#[test]
+fn check_module_graph_entry_does_not_seed_private_methods_for_imported_types() {
+    let tmp = tempfile::tempdir().expect("temp dir");
+    let geometry_path = tmp.path().join("geometry.zen");
+    std::fs::write(
+        &geometry_path,
+        r#"pub Point: { x: i32 }
+
+Point.value = (self: Point) i32 {
+    self.x
+}
+"#,
+    )
+    .expect("write imported module");
+
+    let main_path = tmp.path().join("main.zen");
+    std::fs::write(
+        &main_path,
+        r#"{ Point } = geometry
+
+main = () i32 {
+    point = Point { x: 7 }
+    point.value()
+}
+"#,
+    )
+    .expect("write entry module");
+
+    let mut files = crate::error::FileTable::new();
+    let mut modules = crate::module_system::ModuleSystem::new();
+    let graph = modules
+        .load_module_graph(&main_path, &mut files)
+        .expect("module graph");
+
+    let err = TypeChecker::new()
+        .check_module_graph_entry(&graph)
+        .expect_err("private imported methods should not be seeded");
+
+    assert!(
+        err.iter()
+            .any(|d| d.message.contains("type `Point` has no method `value`")),
+        "expected private imported method diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_module_graph_entry_specializes_public_generic_methods_for_imported_types() {
+    let tmp = tempfile::tempdir().expect("temp dir");
+    let geometry_path = tmp.path().join("geometry.zen");
+    std::fs::write(
+        &geometry_path,
+        r#"pub Point: { x: i32 }
+
+pub Point.keep<T> = (self: Point, value: T) T {
+    value
+}
+"#,
+    )
+    .expect("write imported module");
+
+    let main_path = tmp.path().join("main.zen");
+    std::fs::write(
+        &main_path,
+        r#"{ Point } = geometry
+
+main = () i32 {
+    point = Point { x: 7 }
+    point.keep<i32>(1)
+}
+"#,
+    )
+    .expect("write entry module");
+
+    let mut files = crate::error::FileTable::new();
+    let mut modules = crate::module_system::ModuleSystem::new();
+    let graph = modules
+        .load_module_graph(&main_path, &mut files)
+        .expect("module graph");
+
+    let mut tc = TypeChecker::new();
+    let typed = tc
+        .check_module_graph_entry(&graph)
+        .expect("imported public type should seed public generic method templates");
+
+    assert!(typed
+        .functions
+        .iter()
+        .any(|function| function.name == "Point.keep_i32"));
+}
